requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
pytest>=8.0.0
httpx>=0.27.0
